Move Shape queries and projections specific to the Spectacle interface to their own modules
