+++ workspaces/optic-engine/src/projections/spectacle/mod.rs
@@ -1,1 +1,2 @@
 pub mod endpoints;
+pub mod shape;
