+++ workspaces/optic-engine/src/queries/mod.rs
@@ -1,6 +1,7 @@
 pub mod endpoint;
 pub mod history;
 pub mod shape;
+pub mod spectacle;
 
 pub use endpoint::EndpointQueries;
 pub use history::HistoryQueries;
