+++ workspaces/optic-engine/src/queries/spectacle/mod.rs
@@ -1,0 +1,3 @@
+pub mod spec_choices;
+
+pub use spec_choices::*;
